// src/app/model/vehicule.ts
export class Vehicule {

  idVehicule?: number;
  model?: string;
  registrationNumber?: string;
  image?: string;
  type?: string;
  color?: string;
  idSupplier?: string;
  statusVehicule?: string;
  location?: string;
  idLivreur?: number;
  createdAt?: Date;
  createdBy?: string;
  updatedAt?: Date;
  updatedBy?: string;
}